JNG-2099: refactor transformation to handle double type col and row values

+++ judo-tatami-esm2ui/src/test/java/hu/blackbelt/judo/tatami/esm2ui/Esm2UiPreviewTest.java
@@ -339,8 +339,8 @@ public class Esm2UiPreviewTest {
         
         if(UiUtils.isHorizontalFlex(element)) {
           sb.append("---HORIZONTAL---");
-          int sumCol = UiUtils.getSumOfCols((hu.blackbelt.judo.meta.ui.Container) element);
-          int maxRow = UiUtils.getMaxOfRows((hu.blackbelt.judo.meta.ui.Container) element);
+          double sumCol = UiUtils.getSumOfCols((hu.blackbelt.judo.meta.ui.Container) element);
+          double maxRow = UiUtils.getMaxOfRows((hu.blackbelt.judo.meta.ui.Container) element);
           sb.append(maxRow + " rows in " + element.getRow());
           if (maxRow <= element.getRow()) sb.append(" ***FITS*** //////// ");
           else sb.append(" ===WRONG=== //////// ");
@@ -350,8 +350,8 @@ public class Esm2UiPreviewTest {
         }
         else if(UiUtils.isVerticalFlex(element)) {
           sb.append("|||VERTICAL||| ");
-          int sumRow = UiUtils.getSumOfRows((hu.blackbelt.judo.meta.ui.Container) element);
-          int maxCol = UiUtils.getMaxOfCols((hu.blackbelt.judo.meta.ui.Container) element);
+          double sumRow = UiUtils.getSumOfRows((hu.blackbelt.judo.meta.ui.Container) element);
+          double maxCol = UiUtils.getMaxOfCols((hu.blackbelt.judo.meta.ui.Container) element);
 
           sb.append(sumRow + " rows in " + element.getRow());
           if (sumRow <= element.getRow()) sb.append(" ***FITS*** //////// ");
